move CORS configuration to a separate module
